pyproject.toml
[project]
name = "flowr"
version = "0.2.20260427"
description = "non-deterministic state machine specification to knead workflows"
readme = "README.md"
requires-python = ">=3.13"
license = { file = "LICENSE" }
authors = [
    { name = "eol", email = "nullhack@users.noreply.github.com" }
]
maintainers = [
    { name = "eol", email = "nullhack@users.noreply.github.com" }
]
dependencies = [
    "pyyaml>=6.0",
]

[project.urls]
Repository = "https://github.com/nullhack/flowr"
Documentation = "https://github.com/nullhack/flowr/tree/main/docs/api/"

[project.optional-dependencies]
dev = [
    "pdoc>=14.0",
    "pytest>=9.0.3",
    "pytest-html>=4.0",
    "pytest-cov>=6.1.1",
    "pytest-mock>=3.14.0",
    "ruff>=0.11.5",
    "taskipy>=1.14.1",
    "hypothesis>=6.148.4",
    "pyright>=1.1.407",
    "ghp-import>=2.1.0",
]

[tool.setuptools]
packages = ["flowr"]

[tool.ruff.lint]
ignore = []
select = [
  "A",
  "ANN",
  "ASYNC",
  "B",
  "C4",
  "C9",
  "D",
  "DTZ",
  "E",
  "ERA",
  "F",
  "FURB",
  "G",
  "I",
  "ICN",
  "LOG",
  "N",
  "NPY",
  "PD",
  "PT",
  "PTH",
  "R",
  "RUF",
  "S",
  "SIM",
  "T20",
  "TD",
  "W",
]
preview = true
mccabe.max-complexity = 10
pydocstyle.convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S404", "ANN", "D205", "D212", "D415", "D100", "D101", "D102", "D103"]
"scripts/*.py" = ["T20", "S311"]

[tool.pytest.ini_options]
minversion = "6.0"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "deprecated: marks tests for deprecated AC; automatically skipped (deselect with '-m \"not deprecated\"')",
    "bug: marks tests that reproduce a reported defect (deselect with '-m \"not bug\"')",
]
addopts = """
--maxfail=10 \
--color=yes \
--tb=short \
-q \
"""
testpaths = ["tests"]
python_files = ["*_test.py"]
python_functions = ["test_*"]

[tool.coverage.report]
fail_under = 100
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "if 0:",
    "if __name__ == .__main__.:",
]

[tool.taskipy.tasks]
run = "python -m flowr"
test-coverage = """\
pytest \
  --cov-config=pyproject.toml \
  --cov=flowr \
  --tb=no
"""
test-build = """\
pytest \
  --doctest-modules \
  --cov-config=pyproject.toml \
  --cov-report html:docs/coverage \
  --cov-report term:skip-covered \
  --cov=flowr \
  --hypothesis-show-statistics \
  --html=docs/tests/report.html \
  --self-contained-html \
"""
test = "pytest --tb=short"
test-fast = "pytest -m \"not slow\" -q --no-header --tb=no"
test-slow = "pytest -m slow"
ruff-check = "ruff check . --fix"
ruff-format = "ruff format ."
ruff-format-check = "ruff format . --check"
lint = "task ruff-check && task ruff-format"
doc-serve = "pdoc ./flowr --host localhost --port 8080"
doc-build = """\
pdoc ./flowr -o docs/api --search && \
pytest \
  --cov-config=pyproject.toml \
  --cov-report html:docs/coverage \
  --cov=flowr \
  --html=docs/tests/report.html \
  --self-contained-html \
  -q \
"""
doc-publish = "task doc-build && ghp-import -n -p -f docs"
static-check = "pyright flowr tests"
assign-ids = "python scripts/assign_ids.py"
release-check = "python scripts/check_version.py && task lint && task static-check && task test && task doc-build"

[dependency-groups]
dev = [
    "safety>=3.7.0",
]


